move EventTriggers component to core-ui/shared
