+++ core-ui/src/components/Lambdas/helpers/misc/index.js
@@ -1,10 +1,8 @@
 export * from './backendModulesExist';
-export * from './bemClasses';
 export * from './constants';
 export * from './formatMessage';
 export * from './handleSubscriptionEvent';
 export * from './omitTypenames';
 export * from './renderMarkdown';
-export * from './searchForObject';
 export * from './useConfigData';
 export * from './useLogsView';
