+++ core-ui/src/shared/components/EventTriggers/CreateEventTriggerForm.scss
