+++ core-ui/src/shared/components/EventTriggers/Schema/Schema.scss
