+++ core-ui/src/shared/components/EventTriggers/Schema/SchemaProperties.js
