+++ core-ui/src/shared/components/EventTriggers/test/CreateEventTriggerModal.test.js
