+++ core-ui/src/shared/helpers/index.js
@@ -1,0 +1,2 @@
+export * from './bemClasses';
+export * from './searchForObject';
