+++ core-ui/src/shared/helpers/searchForObject.js
